+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - 添加了使用 Github 远程模板的功能
 - 全局设置指令：
 
-  - `hamibot-assistant.choseDefaultTemplate` : 从默认提供的模板中选择
+  - `hamibot-assistant.setProjectTemplate` : 从默认提供的模板中选择
 
 - 添加了创建时使用手动指定模板的功能